fix(oss): tint annotation label-type chips for dark theme [TH-7263]
The chips hardcoded light-theme pastels with no dark variant, so they
rendered as near-white blocks on a dark surface. Extract one theme-aware
chip and drop the duplicate copy that carried the same bug.
